+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - WebGL runtime error for Unity2020 build due to unsupported compression setting.
+- Unity WebGL template path wasn't updating to Zappar2020 on `Zappar/Editor/UpdateProjectSettingsToPublish`.
 
 
 ## [1.1.0-preview.2] - 2021-07-05